+++ apis/test/influxdb2/users_api_test.rb
@@ -1,0 +1,44 @@
+# The MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+require 'test_helper'
+
+class UsersApiTest < BaseApiTests
+  def setup
+    super
+    @client.create_users_api.get_users.users.each do |user|
+      next unless user.name.end_with?('_TEST')
+
+      @client.create_users_api.delete_users_id(user.id)
+    end
+  end
+
+  def test_create_user
+    name = generate_name('user')
+    user = InfluxDB2::API::User.new(name: name)
+
+    result = @client.create_users_api.post_users(user)
+
+    refute_nil result.id
+    refute_nil result.links
+    assert_equal name, result.name
+    assert_equal 'active', result.status
+  end
+end